# InitializrSpringbootProjectFresh/src/main/resources/api-reference/merchant-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cito Gateway integration and merchant API reference",
    "version": "2.0",
    "description": "Authenticated integration and merchant workspace APIs. Session operations require the merchant portal; they are not API-key integrations. API access is metered separately from service charges. See the integration guide and current portal rates."
  },
  "servers": [
    {
      "url": "https://YOUR-CITO-HOST",
      "description": "Replace with the approved environment URL supplied by Cito."
    }
  ],
  "paths": {
...
    "/api/v2/channels": {
      "get": {
        "summary": "List supported payment channels",
        "operationId": "listChannels",
        "description": "Requires a verified merchant v2 signature. API access rate applies per authenticated request.",
        "parameters": [
          {
            "name": "merchantNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supported channels"
          }
        },
        "security": [
          {
            "cpay_494ca4_CPayV2Signature": []
          }
        ],
        "x-cito-audience": "THIRD_PARTY",
        "x-cito-billing": {
          "unit": "AUTHENTICATED_REQUEST",
          "defaultRate": "0.0000",
          "currency": "UGX",
          "currentRates": "/api/v2/portal/api-reference/rates"
        },
        "x-source-contract": "cpay-v2-openapi.yaml"
      }
    },
...
    "/api/v2/payments/{reference}": {
      "get": {
        "summary": "Check transaction status",
        "operationId": "transactionStatus",
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "merchantNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction status"
          },
          "401": {
            "$ref": "#/components/responses/cpay_494ca4_Unauthorized"
          }
        },
        "security": [
          {
            "cpay_494ca4_CPayV2Signature": []
          }
        ],
        "x-cito-audience": "THIRD_PARTY",
        "x-cito-billing": {
          "unit": "AUTHENTICATED_REQUEST",
          "defaultRate": "0.0000",
          "currency": "UGX",
          "currentRates": "/api/v2/portal/api-reference/rates"
        },
        "x-source-contract": "cpay-v2-openapi.yaml"
      }
    },
    "/api/v2/accounts/validate": {
      "post": {
        "summary": "Validate an account name before payout",
        "operationId": "validateAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cpay_494ca4_AccountValidationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cpay_494ca4_AccountValidationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/cpay_494ca4_BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/cpay_494ca4_Unauthorized"
          }
        },
        "security": [
          {
            "cpay_494ca4_CPayV2Signature": []
          }
        ],
        "x-cito-audience": "THIRD_PARTY",
        "x-cito-billing": {
          "unit": "AUTHENTICATED_REQUEST",
          "defaultRate": "0.0000",
          "currency": "UGX",
          "currentRates": "/api/v2/portal/api-reference/rates"
        },
        "x-source-contract": "cpay-v2-openapi.yaml"
      }
    },
...
    "/api/v2/statements": {
      "get": {
        "summary": "Export merchant statement rows",
        "operationId": "exportMerchantStatements",
        "parameters": [
          {
            "name": "merchantNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ],
              "default": "json"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Merchant statement export",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cpay_494ca4_StatementExportResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/cpay_494ca4_BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/cpay_494ca4_Unauthorized"
          }
        },
        "security": [
          {
            "cpay_494ca4_CPayV2Signature": []
          }
        ],
        "x-cito-audience": "THIRD_PARTY",
        "x-cito-billing": {
          "unit": "AUTHENTICATED_REQUEST",
          "defaultRate": "0.0000",
          "currency": "UGX",
          "currentRates": "/api/v2/portal/api-reference/rates"
        },
        "x-source-contract": "cpay-v2-openapi.yaml"
      }
    },
    "/api/v2/payments/collect": {
      "post": {
        "summary": "Submit a mobile money collection request (compat orchestration path)",
        "operationId": "collectPayment",
        "parameters": [
          {
            "$ref": "#/components/parameters/cpay_494ca4_IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cpay_494ca4_PaymentRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Payment accepted for processing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cpay_494ca4_PaymentResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/cpay_494ca4_BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/cpay_494ca4_Unauthorized"
          }
        },
        "security": [
          {
            "cpay_494ca4_CPayV2Signature": []
          }
        ],
        "x-cito-audience": "THIRD_PARTY",
        "x-cito-billing": {
          "unit": "AUTHENTICATED_REQUEST",
          "defaultRate": "0.0000",
          "currency": "UGX",
          "currentRates": "/api/v2/portal/api-reference/rates"
        },
        "x-source-contract": "cpay-v2-openapi.yaml"
      }
    },
    "/api/v2/payments/payout": {
      "post": {
        "summary": "Submit a mobile money payout request (compat orchestration path)",
        "operationId": "payoutPayment",
        "parameters": [
          {
            "$ref": "#/components/parameters/cpay_494ca4_IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cpay_494ca4_PaymentRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Payout accepted; may be APPROVAL_PENDING if a maker-checker limit fired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cpay_494ca4_PaymentResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/cpay_494ca4_BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/cpay_494ca4_Unauthorized"
          }
        },
        "security": [
          {
            "cpay_494ca4_CPayV2Signature": []
          }
        ],
        "x-cito-audience": "THIRD_PARTY",
        "x-cito-billing": {
          "unit": "AUTHENTICATED_REQUEST",
          "defaultRate": "0.0000",
          "currency": "UGX",
          "currentRates": "/api/v2/portal/api-reference/rates"
        },
        "x-source-contract": "cpay-v2-openapi.yaml"
      }
    },
    "/api/v2/balances": {
      "get": {
        "summary": "List a merchant's available channel balances",
        "operationId": "merchantBalances",
        "parameters": [
          {
            "name": "merchantNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Channel balances"
          },
          "400": {
            "$ref": "#/components/responses/cpay_494ca4_BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/cpay_494ca4_Unauthorized"
          }
        },
        "security": [
          {
            "cpay_494ca4_CPayV2Signature": []
          }
        ],
        "x-cito-audience": "THIRD_PARTY",
        "x-cito-billing": {
          "unit": "AUTHENTICATED_REQUEST",
          "defaultRate": "0.0000",
          "currency": "UGX",
          "currentRates": "/api/v2/portal/api-reference/rates"
        },
        "x-source-contract": "cpay-v2-openapi.yaml"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cpay_494ca4_CPayV2Signature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CPay-Signature"
      },
      "cito_8a569b_CPayV2Signature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CPay-Signature",
        "description": "Signature-based merchant authentication enforced by the CPay/Cito v2 request security service."
      },
      "cito_8a569b_MerchantSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID",
        "description": "Session established by Cito sign-in. Entitlements are enforced per merchant, service and environment."
      },
      "cito_a767bf_signedMerchantRequest": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CPay-Signature",
        "description": "Cito signed merchant-request authentication. Merchant and timestamp headers are validated by the shared v2 request security layer."
      },
      "cito_a767bf_merchantSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      },
      "cito_a12dea_CitoBaasApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Cito-Api-Key"
      },
      "cito_69f952_merchantSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "SESSION"
      }
    },
    "parameters": {
      "cpay_494ca4_IdempotencyKey": {
        "name": "X-CPay-Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "Merchant-scoped request identity. Identical requests replay their stored result; different bodies conflict. Failed request cleanup replays REQUEST_FAILED without releasing the key. Abandoned claims recover after 15 minutes. Check the payment reference before using a new key; REQUEST_FAILED does not prove provider failure.",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
      "cpay_494ca4_PaymentParty": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "cpay_494ca4_PaymentRequest": {
        "type": "object",
        "required": [
          "merchantNumber",
          "country",
          "currency",
          "amount",
          "reference",
          "description",
          "callbackUrl"
        ],
        "properties": {
          "merchantNumber": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "callbackUrl": {
            "type": "string"
          },
          "payer": {
            "$ref": "#/components/schemas/cpay_494ca4_PaymentParty"
          },
          "payee": {
            "$ref": "#/components/schemas/cpay_494ca4_PaymentParty"
          }
        }
      },
      "cpay_494ca4_PaymentResult": {
        "type": "object",
        "description": "Payment lifecycle result. REQUEST_FAILED represents an incomplete API request, not a terminal provider outcome; query the reference before retrying.",
        "properties": {
          "reference": {
            "type": "string"
          },
          "transactionId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "providerResponse": {
            "type": "string"
          }
        }
      },
...
      "cpay_494ca4_AccountValidationRequest": {
        "type": "object",
        "required": [
          "merchantNumber",
          "msisdn"
        ],
        "properties": {
          "merchantNumber": {
            "type": "string"
          },
          "msisdn": {
            "type": "string"
          }
        }
      },
      "cpay_494ca4_AccountValidationResponse": {
        "type": "object",
        "properties": {
          "msisdn": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "cpay_494ca4_StatementExportResponse": {
        "type": "object",
        "properties": {
          "merchantNumber": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "count": {
            "type": "integer"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/cpay_494ca4_StatementRow"
            }
          }
        }
      },
      "cpay_494ca4_StatementRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdOn": {
            "type": "string"
          },
          "gatewayId": {
            "type": "string"
          },
          "transactionType": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "merchantReference": {
            "type": "string"
          },
          "transactionId": {
            "type": "string"
          },
          "transactionStatus": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "cpay_494ca4_BadRequest": {
        "description": "Request was rejected"
      },
      "cpay_494ca4_Unauthorized": {
        "description": "Request signature or credentials were rejected"
      },
...
    },
...
  },
...
}